+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/log_attributes.py
@@ -40,6 +40,12 @@ LOG_IOSTREAM: Final = "log.iostream"
 The stream associated with the log. See below for a list of well-known values.
 """
 
+LOG_RECORD_ORIGINAL: Final = "log.record.original"
+"""
+The complete orignal Log Record.
+Note: This value MAY be added when processing a Log Record which was originally transmitted as a string or equivalent data type AND the Body field of the Log Record does not contain the same value. (e.g. a syslog or a log record read from a file.).
+"""
+
 LOG_RECORD_UID: Final = "log.record.uid"
 """
 A unique identifier for the Log Record.