+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/pool_attributes.py
@@ -16,5 +16,5 @@ from typing import Final
 
 POOL_NAME: Final = "pool.name"
 """
-Deprecated: Replaced by `db.client.connections.pool.name`.
+Deprecated: Replaced by `db.client.connection.pool.name`.
 """